Got as far as PARAMETERS on the documentation of variables.
Noted that starting values dictate ending values, so we need to confirm algorithm is working right.

There is a single well pumping, but there is so little variability in the observations that I think
the K-field is super nonunique. Perhaps we need to add recharge. This may not be the best test
problem. Something to chat with Marco about....