Added posterior covariance module. Updated documentation and variable list.
The 'posterior_cov_flag' must to be added in algorithmic_cv block of the
.bgp file to activate the posterior covariance evaluation.
In the main I have indicated with VV the full posterior covariance matrix
and with V its diagonal (calculated only in case of compression of Q).
Only VV or V is allocaed based on the case.

Mike I think there is an error in the BPO_OUTPUT_CONTROL file, line 212-216.
In fact there is a warning when you run the code.
There is a "do" loop with index "i", but the following variables don't
have indices. I think should be Q0_All(i)%..... Take a look please.

Now I will work on the Qthetatheta stuff. I hope to finish before lunch!