Added monitoring of the objective function during structural parameter optimization.
Actually right now we can use both:
objective function and norm of differences between current
and previous structural parameters.

It works in this way: a positive value of structural_conv in the .bgp file
means to monitore the objective function, a negative value means
to monitor the norm of differences between parameters (of course we take
the absolute value of structural_conv in this case).

If we like this way I will update the documentation too.

I added a subroutine that in case of objective function monitoring,
at the first bga loop, calculates the value of the objective function
with the initial structural parameters (easy to calculate because all the
variable are already based on these structural parameters and dtheta is zero).
I did this to compare at the first bga iteration the new estimated
objective function with the one calculated with the initial structural parameters.

Mike, let me know what you think.
If you have problem to understand what the variables are when you will
write the code to report the structural parameter values let me know.
I tried to be clear into the code adding some comment.