Fixed bug in printing out posterior covariance matrix or vector.
The problem was in bpo_output_control file at the subroutine
bpo_write_allpars_95ci.
There was here the variable SD(:) that was representing V(:) and at a
certain point there was the calculation of sqrt(SD).
SD was not a local variable so on exit from the subroutine also V was
updated with the sqrt(V). This was the reason of the differences you found.
I modified a bit the subroutine to solve this problem and to deallocate
the other vectors no more necessary.

Moreover I modified the MAIN a bit in calling the subroutine to write
posterior covariance.

Probably we need to say in documentation,
in which space is the posterior covariance.
It is now in the estimation space in case of log transformation, right?

Still remains to printing out the structural parameters.