Added output of structural parameters.
Also worked on documentation, including changing the default for theta
transformation from 1 to 0. Erasing comments in the LyX file as they
are addressed. Added figures for Toeplitz and a very general flowchart
to help define inner and outer iterations. Need to look at the logic
of the flowchart to see if linesearch and structural parameters
should be listed inside the loop of inner iterations or outside of it
as they are now.

But....now it's time for beer!