Solved problem on structural parameter optimization.
Mike, the problem was in the MAIN. After writing the intermediate
structural parameter vector (just the parameters we optimize for)
there was the line: cv_S%struct_par_opt = 0 (and cv_S%struct_par_opt = 1
before the writing line). In this way after the first iteration the
structural parameters were no more estimated and the estimation was
continuing just for sigma, giving big values and the overfitting problem
you found. The overfitting I think was correct because the alghorithm
tried to find the minimum of the structural objective function just
modifying sigma and holding the other parameters on wrong values.

I have correct this and moreover I modified a bit the writing routine.
Now the code writes, at each iteration, the values of the structural
parameters we ask to optimizing for and at the end the final structural
parameters (all of them, including also the structural parameters we
didn't optimize for).

I hope the problem was just here! If you run now the three layer model
you should obtain a really good solution. Let me know.

Marco