Solved the stack oveflow problems in case of many observations.
I modified also the POINTERS (that were not into TYPE variables)
in ALLOCATABLE variables. These solved many memory problems.
I run some tests and the results are identically to those obtained
with the old version of the code.
My model with 600 observations now run without memory overflows.

One more thingh Mike:
I saw that in the .bpr file, at the beginnig, when we write out the
structural parameter control variables, for the epistemic uncertinty
we do not report the flag that says if we are optimizing or not for
sigma and also if the power transformation is active for sigma and
eventually the alpha value.
Can you take care of this when you have time?

Marco

PS: I compiled also a 64bit version of bgaPEST
and it seems faster.